MH3-30049: Regenerated all MPLABX based applications

+++ apps/aria_mxt_configure/firmware/src/config/lcc_rgb565_mxt/gfx/hal/inc/gfx_display.h
@@ -74,16 +74,16 @@ typedef struct GFX_DisplayInfo_t
         
         struct horizontal_t
         {
-            int8_t pulse_width;
-            int8_t back_porch;
-            int8_t front_porch;
+            int32_t pulse_width;
+            int32_t back_porch;
+            int32_t front_porch;
         } horz;
         
         struct vertical_t
         {
-            int8_t pulse_width;
-            int8_t back_porch;
-            int8_t front_porch;
+            int32_t pulse_width;
+            int32_t back_porch;
+            int32_t front_porch;
         } vert;
         
         int32_t inv_left_shift;
