+++ apps/legato_benchmark/firmware/src/config/ili9488_rgb565_e54_cult_parallel/gfx/legato/common/legato_common.h
@@ -22,7 +22,7 @@
 
 //DOM-IGNORE-BEGIN
 /*******************************************************************************
-* Copyright (C) 2018 Microchip Technology Inc. and its subsidiaries.
+* Copyright (C) 2020 Microchip Technology Inc. and its subsidiaries.
 *
 * Subject to your compliance with these terms, you may use Microchip software
 * and any derivatives exclusively with Microchip products. It is your
@@ -323,10 +323,10 @@ static const leSize leSize_Zero = {0, 0};
 */
 typedef struct leRect
 {
-    int32_t x;
-    int32_t y;
-    int32_t width;
-    int32_t height;
+    int16_t x;
+    int16_t y;
+    int16_t width;
+    int16_t height;
 } leRect;
 
 typedef void* leBuffer;
